MCP_agent_communication/requirements.txt
pydantic-ai==1.0.0
treequest==0.1.2
langgraph>=0.2,<1
langchain-openai>=0.2
python-dotenv>=1.0
langchain_community==0.3.29
langchain_experimental==0.3.4
mcp-run-python==0.0.21



